requirements.txt
python-telegram-bot>=20.0
instagrapi>=1.16.30
